splitter GrunnlagPensjonsavtaler og skriver tester
