+++ src/components/Grunnlag/GrunnlagPensjonsavtaler/__tests__/GrunnlagPensjonsavtaler.test.tsx
@@ -9,7 +9,7 @@ import {
   userInputInitialState,
   Simulation,
 } from '@/state/userInput/userInputReducer'
-import { render, screen, userEvent, waitFor } from '@/test-utils'
+import { render, screen, userEvent } from '@/test-utils'
 
 describe('GrunnlagPensjonsavtaler', () => {
   const fakeInntektApiCall = {
@@ -242,165 +242,6 @@ describe('GrunnlagPensjonsavtaler', () => {
           exact: false,
         })
       ).toBeVisible()
-    })
-
-    it('rendrer riktig med avtaler som bare har start dato', async () => {
-      const user = userEvent.setup()
-      const avtale: Pensjonsavtale = {
-        key: 0,
-        produktbetegnelse: 'DNB',
-        kategori: 'PRIVAT_TJENESTEPENSJON',
-        startAar: 67,
-        utbetalingsperioder: [
-          {
-            startAlder: { aar: 67, maaneder: 0 },
-            aarligUtbetaling: 12345,
-            grad: 100,
-          },
-        ],
-      }
-      mockResponse('/v1/pensjonsavtaler', {
-        status: 200,
-        json: {
-          avtaler: [
-            avtale,
-            {
-              ...avtale,
-              utbetalingsperioder: [
-                {
-                  ...avtale.utbetalingsperioder[0],
-                  startAlder: { aar: 67, maaneder: 6 },
-                },
-              ],
-            },
-          ],
-          utilgjengeligeSelskap: [],
-        },
-        method: 'post',
-      })
-      const { container } = render(<GrunnlagPensjonsavtaler />, {
-        preloadedState: {
-          /* eslint-disable @typescript-eslint/ban-ts-comment */
-          // @ts-ignore
-          api: { ...fakeInntektApiCall },
-          userInput: {
-            ...userInputInitialState,
-            samtykke: true,
-            currentSimulation: currentSimulation,
-          },
-        },
-      })
-      const buttons = screen.getAllByRole('button')
-      await user.click(buttons[0])
-      expect(await screen.findByTestId('pensjonsavtaler-table')).toBeVisible()
-      expect(
-        await screen.findByText('grunnlag.pensjonsavtaler.tabell.title.left')
-      ).toBeVisible()
-      expect(
-        await screen.findByText('grunnlag.pensjonsavtaler.tabell.title.right')
-      ).toBeVisible()
-      await waitFor(async () => {
-        expect(await screen.findByText('Privat tjenestepensjon')).toBeVisible()
-        expect(
-          await screen.findByText(
-            'grunnlag.pensjonsavtaler.livsvarig 67 grunnlag.pensjonsavtaler.aar'
-          )
-        ).toBeVisible()
-        expect(
-          await screen.findByText(
-            'grunnlag.pensjonsavtaler.livsvarig 67 grunnlag.pensjonsavtaler.aar grunnlag.pensjonsavtaler.og 6 grunnlag.pensjonsavtaler.md'
-          )
-        ).toBeVisible()
-        expect(await screen.findAllByText('12 345 kr')).toHaveLength(2)
-        const rows = container.querySelectorAll('tr')
-        expect(rows?.length).toBe(6)
-        expect(
-          await screen.findByText(
-            'Alle avtaler i privat sektor er hentet fra',
-            {
-              exact: false,
-            }
-          )
-        ).toBeVisible()
-      })
-    })
-
-    it('rendrer riktig med avtaler som har både start- og sluttdato', async () => {
-      const user = userEvent.setup()
-      const avtale: Pensjonsavtale = {
-        key: 0,
-        produktbetegnelse: 'DNB',
-        kategori: 'PRIVAT_TJENESTEPENSJON',
-        startAar: 67,
-        sluttAar: 77,
-        utbetalingsperioder: [
-          {
-            startAlder: { aar: 67, maaneder: 0 },
-            sluttAlder: { aar: 77, maaneder: 11 },
-            aarligUtbetaling: 12345,
-            grad: 100,
-          },
-          {
-            startAlder: { aar: 67, maaneder: 6 },
-            sluttAlder: { aar: 77, maaneder: 1 },
-            aarligUtbetaling: 12345,
-            grad: 100,
-          },
-        ],
-      }
-      mockResponse('/v1/pensjonsavtaler', {
-        status: 200,
-        json: {
-          avtaler: [avtale],
-          utilgjengeligeSelskap: [],
-        },
-        method: 'post',
-      })
-      const { container } = render(<GrunnlagPensjonsavtaler />, {
-        preloadedState: {
-          /* eslint-disable @typescript-eslint/ban-ts-comment */
-          // @ts-ignore
-          api: { ...fakeInntektApiCall },
-          userInput: {
-            ...userInputInitialState,
-            samtykke: true,
-            currentSimulation: currentSimulation,
-          },
-        },
-      })
-      const buttons = screen.getAllByRole('button')
-      await user.click(buttons[0])
-      expect(await screen.findByTestId('pensjonsavtaler-table')).toBeVisible()
-      expect(
-        await screen.findByText('grunnlag.pensjonsavtaler.tabell.title.left')
-      ).toBeVisible()
-      expect(
-        await screen.findByText('grunnlag.pensjonsavtaler.tabell.title.right')
-      ).toBeVisible()
-      await waitFor(async () => {
-        expect(await screen.findByText('Privat tjenestepensjon')).toBeVisible()
-        expect(
-          await screen.findByText(
-            'grunnlag.pensjonsavtaler.fra_og_med 67 grunnlag.pensjonsavtaler.aar grunnlag.pensjonsavtaler.til_og_med 77 grunnlag.pensjonsavtaler.aar'
-          )
-        ).toBeVisible()
-        expect(
-          await screen.findByText(
-            'grunnlag.pensjonsavtaler.fra_og_med 67 grunnlag.pensjonsavtaler.aar grunnlag.pensjonsavtaler.og 6 grunnlag.pensjonsavtaler.md grunnlag.pensjonsavtaler.til_og_med 77 grunnlag.pensjonsavtaler.aar grunnlag.pensjonsavtaler.og 1 grunnlag.pensjonsavtaler.md'
-          )
-        ).toBeVisible()
-        expect(await screen.findAllByText('12 345 kr')).toHaveLength(2)
-        const rows = container.querySelectorAll('tr')
-        expect(rows?.length).toBe(5)
-        expect(
-          await screen.findByText(
-            'Alle avtaler i privat sektor er hentet fra',
-            {
-              exact: false,
-            }
-          )
-        ).toBeVisible()
-      })
     })
   })
 })
